Particles
I added dash, uppercut, heal, fire and blood as particles for the
player. There is a large fire that can be used for environment. Bubble
particle for the sink has been improved. The dash has 4 different
colours (Red, Yellow, Blue & Green) meant to belong to different
players. All the particles are put in the Prefab folder. Materials is in
Materials - Particles